Add files via upload
The objective of the game is to help the space dog, Dot, hit the ball as many times as possible. You navigate Dot using the arrow keys on your keyboard. The ball will move around randomly, which is easy enough, but you also have to make sure Dot doesn't hit the dragonfly. If Dot hits the dragonfly, the animations stop and the game is over. You can hit the green flag to start over; the score counter will reset to 0.